+++ src/services/spaceService.js
@@ -22,25 +22,39 @@
  */
 
 import axios from '@nextcloud/axios'
-import { ESPACE_GID_PREFIX, ESPACE_MANAGERS_PREFIX, ESPACE_USERS_PREFIX } from '../constants'
+import { ESPACE_GID_PREFIX, ESPACE_MANAGERS_PREFIX, ESPACE_USERS_PREFIX } from '../constants.js'
 import { generateUrl } from '@nextcloud/router'
+import BadCreateError from '../Errors/BadCreateError.js'
+import NotificationError from './Notifications/NotificationError.js'
 
 /**
 	* @param {string} spaceName it's a name for the space to create
 	* @param {number} folderId it's the id of groupfolder
+	* @param {object} vueInstance it's an instance of vue
 	* @return {object}
 	*/
-export function createSpace(spaceName, folderId) {
+export function createSpace(spaceName, folderId, vueInstance = undefined) {
 	const result = axios.post(generateUrl('/apps/workspace/spaces'),
 		{
 			spaceName,
 			folderId,
 		})
 		.then(resp => {
+			if (typeof (resp.data) !== 'object') {
+				throw new Error('Error when creating a space, it\'s not an object type.')
+			}
+
 			return resp.data
 		})
 		.catch(error => {
-			console.error('createSpace error', error)
+			if (typeof (vueInstance) !== 'undefined') {
+				const toastErrorToCreateWorkspace = new NotificationError(vueInstance)
+				toastErrorToCreateWorkspace.push({
+					title: t('workspace', 'Error to create a workspace'),
+					text: t('workspace', error.message),
+				})
+			}
+			throw new BadCreateError(error.message)
 		})
 	return result
 }
@@ -54,28 +68,27 @@ export function deleteBlankSpacename(spacename) {
 }
 
 /**
-* @param {string} spaceName
+* @param {string} spaceId
 * @param {object} groupfolder
 * @return {object}
 */
-export function convertGroupfolderToSpace(spaceName, groupfolder) {
-	const result = axios.post(generateUrl('/apps/workspace/spaces/convert'),
+export function transferUsersToUserGroup(spaceId, groupfolder) {
+	const result = axios.post(generateUrl(`/apps/workspace/spaces/${spaceId}/transfer-users`),
 		{
-			spaceName,
 			groupfolder,
 		})
 		.then(resp => {
 			return resp.data
 		})
 		.catch(error => {
-			console.error('createSpace error', error)
+			console.error('Error to transfer users', error)
 		})
 	return result
 }
 
 /**
 * @param {string} group it's the groupname to test
-* @return {string}
+* @return {boolean}
 */
 export function isSpaceManagers(group) {
 	const SPACE_MANAGER_REGEX = new RegExp('^' + ESPACE_GID_PREFIX + ESPACE_MANAGERS_PREFIX)
@@ -84,7 +97,7 @@ export function isSpaceManagers(group) {
 
 /**
 	* @param {string} group it's the groupname to test
-	* @return {string}
+	* @return {boolean}
 	*/
 export function isSpaceUsers(group) {
 	const SPACE_USER_REGEX = new RegExp('^' + ESPACE_GID_PREFIX + ESPACE_USERS_PREFIX)