+++ CHANGELOG.md
@@ -6,9 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2022-12-29
+
+### Fixed
+
+- Fix the deletion of a group after groupfolder converting to workspace ( [#644](https://github.com/arawa/workspace/pull/644) )
+
+## [1.3.0] - 2022-12-23
+
 ### Added
 
 - Add Czech localization([#541](https://github.com/arawa/workspace/pull/541) & [#542](https://github.com/arawa/workspace/pull/542) )
+- Limiting user search ( [#631](https://github.com/arawa/workspace/pull/631) )
 
 ### Chore
 
@@ -20,6 +29,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - bump jest from 28.1.0 to 29.0.3 ( [#563](https://github.com/arawa/workspace/pull/563) )
 - bump babel-jest from 28.1.0 to 29.0.3 ( [#564](https://github.com/arawa/workspace/pull/564) )
 - bump @nextcloud/axios from 1.10.0 to 2.0.0 ( [#565](https://github.com/arawa/workspace/pull/565) )
+
+### Changed
+
+- Refactoring the creation workspace part ( [#609](https://github.com/arawa/workspace/pull/609) )
+- Refactorised the converting groupfolders to workspace ( [#633](https://github.com/arawa/workspace/pull/633) )
+
+### Style 
+
+- Change the settings button ( [#504](https://github.com/arawa/workspace/pull/504) )
+
+## [1.2.3] - 2022-10-28
+
+### Changed
+
+- Comment the convert feature ( [#604](https://github.com/arawa/workspace/pull/604) )
 
 ## [1.2.2] - 2022-08-19
 
@@ -63,7 +87,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Define users' roles (GeneralManager and WorkspacesManager)
 
 
-[Unreleased]: https://github.com/arawa/workspace/compare/v1.2.2...main
+[Unreleased]: https://github.com/arawa/workspace/compare/v1.3.1...main
+[1.3.1]: https://github.com/arawa/workspace/compare/v1.3.0...1.3.1
+[1.3.0]: https://github.com/arawa/workspace/compare/v1.2.3...1.3.0
+[1.2.3]: https://github.com/arawa/workspace/compare/v1.2.2...1.2.3
 [1.2.2]: https://github.com/arawa/workspace/compare/v1.2.1...1.2.2
 [1.2.1]: https://github.com/arawa/workspace/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/arawa/workspace/compare/v1.1.0...v1.2.0